requirements.txt
# Framework web
Flask==3.0.2

# Google Calendar API
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Manejo de fechas y zonas horarias
pytz==2024.1
python-dateutil==2.8.2
tzlocal==5.2

# Variables de entorno
python-dotenv==1.0.1

# Servidor WSGI para producción
gunicorn==21.2.0

# Gestión de logs
loguru==0.7.2

# Herramientas de prueba
pytest==7.4.0
pytest-cov==4.1.0 